Add CurrentThreadScheduler
