+++ Scripts/UniTask/Scheduler.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0700fb042f010694782d238049678651
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
