+++ Scripts/UniTask/Scheduler/CurrentThreadScheduler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 21979cc7137a17d4ea8b6202381a02d0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
